Refactor data loading and regions
Refactors the data loading and configuration into provider specific files to reduce the complexity in the main file.

Additionally refactors the defaults and test configuration info by region so that we can start adding additional region configuration options.
