+++ pixi.toml
@@ -14,3 +14,4 @@ ioos_qc = ">=2.1.0,<2.2"
 pandas = ">=2.2.2,<2.3"
 bokeh = "<3"
 pyyaml = ">=6.0.2,<6.1"
+erddapy = ">=2.2.3,<3"
